Exercise 2: MegaBytes Converter
Write a method called printMegyBytesAndKiloBytes that has one parameter type int with the name kiloBytes.
The method should not return anything (void) and it needs to calculate the megabytes and remaining kilobytes from the kilobytes parameter.
If the kilobytes value is less than zero, it needs to print "Invalid Value".
Else it needs to print a message in the format "XX KB = YY MB and ZZ KB".
XX represent the original value of kiloBytes. YY is the calculated megabytes. ZZ is the remaining kilobytes.