Coding Exercise 3: Barking Dog
We have a dog that likes to bark. We need to wake up if the dog is barking at night!
Write a method called shouldWakeUp that has two parameters.
First parameter is boolean barking, represents if the dog is barking. Second parameter is int hourOfDay, represents the hour of the day and has a valid range 0-23.
We have to wake up if the dog is barking before 8 or after 22. In that case return true.
In all other cases return false.