Create Coding Exercise 7: Teen Number Checker
A number is "teen" if it's in the 13-19 range (inclusive).

Write a method hasTeen with three parameters of type int. The method should return boolean and it needs to return true is one of the numbers is "teen". Otherwise, return false.
Write a method isTeen with one parameter type int.  The method should return boolean and it needs to return true if the number is "teen". Otherwise, return false.