Create Coding Exercise 8: Area Calculator
Write a method named area with one double parameter named area. The method needs to return a double value that represents the area of a circle. If the parameter radius is negative, return -1.0 to indicate an invalid value.
Write another, overloaded method with two parameters type double, x and y, that represent the sides of a rectangle. The method needs to return the area of the rectangle. If one or both parameters are negative, return -1.0 to indicate an invalid value.