Create Coding Exercise 9: Minutes To Years And Days Calculator
Write a method called printYearsAndDays with one parameter type long called minutes. The method should not return anything (void) and it needs to calculate years and days from the minutes parameter. If the parameter is less than zero, print text "Invalid Value".
Otherwise, if the parameter is valid, it needs to print the message "XX min = YY y and ZZ d".
XX is the value of minutes, YY is the calculated years, ZZ is the calculated days.