Create Coding Exercise 10: Equality Printer
Write a method printEqual with 3 parameters of type int. The method should not return anything (void). If one of the parameters is less than 0, print "Invalid Value". If all numbers are equal, print "All numbers are equal". If all numbers are different, type "All numbers are different". Otherwise, print "Neither all are equal or different".