Create Coding Exercise 11: Playing Cat
The cats spends most of the day playing (meow!). In particular, they play if the temperature is between 25 and 35 (inclusive). Unless it is summer, then the upper limit is 45 (inclusive) instead of 35.
Write a method called isCatPlaying that has 2 parameters. Method needs to return true if the cat is playing, otherwise return false. The first parameter is type boolean, named summer and represents if it is summer. The second parameter is type int named temperature and represents the temperature.