Create Coding Exercise 12: Number in Word
Write a method called printNumberInWord that has 1 parameter called number which is a whole number. The method needs to print "ZERO", "ONE", "TWO", ..., "NINE", "OTHER" if the int parameter is 0, 1, 2, ..., 9 or other for any other number including negative numbers. You can use switch statement or if-else statement, whatever is easier for you.