Create Coding Exercise 12: Number Of Days In Month
Write a method called isLeapYear with one parameter type int called year. The parameter needs to be greater than 1 and equal to or less than 9999.  If it is not in that range, return false. Otherwise, if it is in the valid range, calculate if it is a leap year and return true if it is, false if it isn't. A leap year is divisible by 4 but not by 100, or it is divisible by 400.
Write another method called getDaysInMonth with two parameters type int, month and year. If month is < 1 or >12, return -1. If year is <1 or >9999, return -1.  The method needs to return the number of days in the month.