Create Coding Exercise 14: Sum Odd
Write a method called isOdd with one parameter type int called number. The method needs to return boolean. If the number is less than 0, return false. If the number is odd, return true, otherwise, return false.
Write a second method called sumOdd with two parameters type int called start and end, which represent a number range.
The method should use a for loop to sum all odd numbers in that range including the end and return the sum. It should call the method isOdd to check if a number is odd. The parameter end needs to be greater than or equal to the parameter start and both start and end parameters need to be greater than 0. Otherwise return -1.