Create Coding Exercise 22: Perfect Number
A perfect number is a positive integer which is equal to the sum of it's divisors. Proper positive divisors are positive divisors that divide the perfect number withour remainder and exclude the perfect number itself.
Write a method isPerfectNumber with one parameter type int called number. If the number is < 1, return false. The method must return true if the number is perfect.